+++ slave/src/src.ino
@@ -40,7 +40,7 @@ void setup(){
 
 		delay(2);
 
-		int flag = radio.listen();
+		unsigned long int flag = radio.listen();
 		radio.phoneHome(&flag, 1);
